+++ src/main/adapters/express/express-route-adapter.ts
@@ -1,6 +1,6 @@
 import { Request, RequestHandler, Response } from 'express'
 
-import { Controller, HttpRequest } from '../../presentation/protocols'
+import { Controller, HttpRequest } from '../../../presentation/protocols'
 
 export const adaptRoute = (controller: Controller): RequestHandler => {
   return async (req: Request, res: Response) => {
